Update to parent pom
Updated to parent pom.  Removed a few of the dependencies inherited from
the parent.  I kept jenkins.version at 1.480, and had to include an
exclusion for sisu-guice.  The lifecycle-mapping configuration stands as
it's less stringent than the parent-pom.   jenkins.version and
github-client  upgrades may be beyond the scope of this work item.