Minor documentation cleanup
Remove the old changelog and link to a tagged version on GitHub.

Replace outdated jenkins-ci.org with jenkins.io domain name.

Link bug reporting to a page that offers bug, improvement, and security
reporting so that users are encouraged to report issues more accurately
and are encouraged to report security issues in the SECURITY project
rather than the JENKINS project.